// server/types/socket.types.ts
import { Card, GameData, Player, PlayerId } from "./types.ts";

export const enum SocketIncoming {
    Join = "join",
    StartGame = "startGame",
    PlayCard = "playCard",
    SelectPlayer = "selectPlayer",
    SelectCard = "selectCard",
    SubmitSelection = "submitSelection",
    AcknowledgeAction = "acknowledgeAction",
}

export const enum SocketOutgoing {
    PlayerUpdate = "playersUpdate",
    Connected = "connected",
    GameUpdate = "gameUpdate",
}

export interface OutgoingGameStateUpdate {
    deckCount: number;
    discard: Array<Card>;
    players: [PlayerId, Player][];
    game: GameData;
}

export interface SocketEvent {
    type: SocketIncoming;
}

export interface StartGameEvent extends SocketEvent {
    type: SocketIncoming.StartGame;
}

export interface JoinEvent extends SocketEvent {
    type: SocketIncoming.Join;
    roomCode: string;
    playerName: string;
}

export interface PlayCardEvent extends SocketEvent {
    type: SocketIncoming.PlayCard;
    cardPlayed: Card;
}

export interface SelectCardEvent extends SocketEvent {
    type: SocketIncoming.SelectCard;
    cardPlayed: Card;
}

export interface SelectPlayerEvent extends SocketEvent {
    type: SocketIncoming.SelectPlayer;
    playerSelected: PlayerId;
}

export interface SubmitSelectionEvent extends SocketEvent {
    type: SocketIncoming.SubmitSelection;
}

export interface AcknowledgeActionEvent extends SocketEvent {
    type: SocketIncoming.AcknowledgeAction;
}

export type SocketMessage =
    | JoinEvent
    | StartGameEvent
    | PlayCardEvent
    | SelectCardEvent
    | SelectPlayerEvent
    | SubmitSelectionEvent
    | AcknowledgeActionEvent;